Refactor damage handling and consolidate ability tests
